Add files via upload
Some weird trouble shooting error..... I emailed you about this and I think there is something off with me using a Mac. I did want you suggested in the email and made sure it is running the correct .csproj file and it will not run with the starter code you provided. It will only run with the .csproj file created when I created the app. I think that is due me using a Mac. I have provided screenshots to show that I have made lots of efforts towards this lab. there is two terminal ss one is before your email and the next is after. Thanks!